refact: 댓글 삭제시 조건 적용, 댓글 - 대댓글이 없을 시 : 바로 삭제 - 대댓글이 있을 시 : comment, userId 변경, isDelete는 true 변경 대댓글 - 바로 삭제 - 부모가 isDelete가 ture일 경우 부모도 삭제

+++ backend/src/main/java/com/project/backend/domain/review/comment/service/ReviewCommentService.java
@@ -76,6 +76,7 @@ public class ReviewCommentService {
                 .comment(reviewCommentDto.getComment())
                 .recommend(new HashSet<>())
                 .depth(0)
+                .isDelete(false)
                 .build();
         if(reviewCommentDto.getParentId()!=null){
             ReviewComment parentComment = reviewCommentRepository.findById(reviewCommentDto.getParentId())
@@ -143,10 +144,28 @@ public class ReviewCommentService {
                         ReviewErrorCode.COMMENT_NOT_FOUND.getMessage()
                 ));
 
-        ReviewCommentDto reviewCommentDto =new ReviewCommentDto(reviewComment);
-        reviewCommentRepository.delete(reviewComment);
+        if(reviewComment.getParent()!=null){
+            ReviewComment parent = reviewComment.getParent();
 
-        return reviewCommentDto;
+            reviewCommentRepository.delete(reviewComment);
+
+            if(parent.isDelete()){
+                reviewCommentRepository.delete(parent);
+            }
+        }else{
+            if(reviewComment.getReplies().size()==0|| reviewComment.getReplies()==null){
+                reviewCommentRepository.delete(reviewComment);
+            }else{
+                reviewComment.setComment("해당 코멘트는 삭제되었습니다. ");
+                reviewComment.setUserId(null);
+                reviewComment.setDelete(true);
+                reviewCommentRepository.save(reviewComment);
+            }
+        }
+
+        return new ReviewCommentDto(reviewComment);
+
+
     }
 
     /**
